Fix 913ef4b18f046239ef3ffcb12997652524fbd35f : Changed uint to unsigned long int for srv_no_of_allowed_nodes. Changed uint to unsigned long int for parameter of srv_bind_to_thread() declaration.

+++ sql/mysqld.h
@@ -138,7 +138,7 @@ extern ulonglong slave_type_conversions_options;
 extern my_bool read_only, opt_readonly;
 extern my_bool srv_numa_enable;
 extern uint srv_no_of_allowed_nodes;
-extern uint srv_allowed_nodes[SRV_MAX_NUM_NUMA_NODES];
+extern unsigned long int srv_allowed_nodes[SRV_MAX_NUM_NUMA_NODES];
 extern MYSQL_PLUGIN_IMPORT my_bool lower_case_file_system;
 extern my_bool opt_enable_named_pipe, opt_sync_frm, opt_allow_suspicious_udfs;
 extern my_bool opt_secure_auth;
